[PATCH] possible issue with connections RT #64718
In using POE::Component::Server::SimpleHTTP v. 2.06, we've come across
a problem with error messages such as:

Can't call method "dead" on an undefined value at
[...]/POE/Component/Server/SimpleHTTP.pm line 840.

The line is:

      # Mark the client dead
      $connection->dead(1);

This happens when our load balancer makes a check to make sure the
port is open where the SimpleHTTP server is listening.

We've fixed it with:

      # Mark the client dead
      $connection->dead(1) if $connection;

I'm afraid at this point I haven't got the ability to do the low-level
debugging necessary to determine exactly how $connection comes to be
undef, or whether this is the best way to accomplish the fix in the
big picture, but I wanted to share the fix in case you found it
useful.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>