post fails if it returns false; check that, not $!
I was getting errors with an ENOENT in them, which dngor says shouldn't be
coming from the kernel.  We discussed this on irc.perl.org #magnet, 2016-12-23
around 15:00 UTC.  One possibility, unconfirmed, is that something else in the
handler is setting $!, and the kernel does not guarantee clearing $! on success
— like lots of Perl 5 stuff.  So, this commit relies on post's return value,
and then assumes $! will be sensible, rather than assuming a true $! means that
the post failed.